Sync latest with v1 tag to use Dockerfile
Updating the `latest` branch to be in sync with the `v1` branch to make sure it is as close to our production setup as possible. 

This PR also brings back the use of the Dockerfile and entrypoint.sh, again to match `v1`. This change was previously made to support running on self-hosted runners. But we have since introduced the `v1-self-hosted` branch which solves that problem.